Remove `innerContainerStyle` and `DetectRender`
innerContainerStyle was used to offset the overlay relative to its
coordinates. Besides having an unhelpful default that anchored the
overlay at "middle-bottom", it seemed to be encouraging people to use
the element for styling purposes. This was never intended. Users should
instead create their own offset element in the content block.

DetectRender is no longer necessary. One less component!